Don't disable turn resolution auto game speed increase on GameSpeedCycle
With `OnlyInTurnResolution` enabled, it's better for the intent of the
option that the automatic game speed increase and resetting doesn't get
disabled on change in ActiveGameSpeed via GameSpeedCycle.

Also, have the GameSpeedCycle in battle with turn resolution affect game
speed only during turn resolution and not when waiting for user input.